@@ -10,60 +10,60 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "^7.1.4",
-    "@angular/common": "^7.1.4",
-    "@angular/compiler": "^7.1.4",
-    "@angular/core": "^7.1.4",
-    "@angular/forms": "^7.1.4",
-    "@angular/http": "^7.1.4",
-    "@angular/platform-browser": "^7.1.4",
-    "@angular/platform-browser-dynamic": "^7.1.4",
-    "@angular/router": "^7.1.4",
+    "@angular/animations": "^8.2.1",
+    "@angular/common": "^8.2.1",
+    "@angular/compiler": "^8.2.1",
+    "@angular/core": "^8.2.1",
+    "@angular/forms": "^8.2.1",
+    "@angular/platform-browser": "^8.2.1",
+    "@angular/platform-browser-dynamic": "^8.2.1",
+    "@angular/router": "^8.2.1",
     "@asymmetrik/ngx-leaflet": "^5.0.1",
-    "@ng-bootstrap/ng-bootstrap": "^4.0.1",
+    "@ng-bootstrap/ng-bootstrap": "^5.1.0",
     "@ng-bootstrap/schematics": "^2.0.0-alpha.1",
     "@ngx-translate/core": "^11.0.1",
     "@ngx-translate/http-loader": "^4.0.0",
+    "ajv": "^6.10.2",
     "bootstrap": "^4.2.1",
     "bootstrap-notify": "^3.1.3",
     "chart.js": "^2.7.3",
-    "core-js": "^2.6.1",
+    "core-js": "^2.6.9",
     "install": "^0.12.2",
     "leaflet": "^1.3.4",
     "left-pad": "^1.3.0",
     "lodash.escaperegexp": "^4.1.2",
-    "messageformat": "^2.0.5",
+    "messageformat": "^2.3.0",
     "moment": "^2.23.0",
-    "ng-recaptcha": "^4.2.1",
+    "ng-recaptcha": "^5.0.0",
     "ng2-charts": "^1.6.0",
     "ng2-file-upload": "^1.3.0",
-    "ngx-autosize": "^1.2.4",
+    "ngx-autosize": "^1.6.3",
     "ngx-moment": "^3.3.0",
-    "ngx-pipes": "^2.3.6",
-    "ngx-translate-messageformat-compiler": "^4.4.0",
-    "npm": "^6.5.0",
+    "ngx-pipes": "^2.5.6",
+    "ngx-translate-messageformat-compiler": "^4.5.0",
+    "npm": "^6.10.3",
     "rxjs": "^6.3.3",
     "sprintf-js": "^1.1.2",
-    "typescript": "^3.1.6 <3.2.0",
-    "zone.js": "^0.8.26"
+    "typescript": "^3.5.3",
+    "zone.js": "^0.9.1"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "^0.11.4",
-    "@angular/cli": "^7.1.4",
-    "@angular/compiler-cli": "^7.1.4",
-    "@angular/language-service": "^7.1.4",
+    "@angular-devkit/build-angular": "^0.802.1",
+    "@angular/cli": "^8.2.1",
+    "@angular/compiler-cli": "^8.2.1",
+    "@angular/language-service": "^8.2.1",
     "@brainly/onesky-utils": "^1.3.1",
-    "@types/jasmine": "^3.3.5",
+    "@types/jasmine": "^3.4.0",
     "@types/jasminewd2": "^2.0.6",
-    "@types/jquery": "^3.3.28",
-    "@types/leaflet": "^1.2.14",
-    "@types/node": "^10.12.18",
-    "codelyzer": "^4.5.0",
+    "@types/jquery": "^3.3.31",
+    "@types/leaflet": "^1.5.0",
+    "@types/node": "^10.14.15",
+    "codelyzer": "^5.1.0",
     "font-awesome": "^4.7.0",
     "jquery": "^3.3.1",
-    "ngx-build-plus": "^7.5.6",
+    "ngx-build-plus": "^8.0",
     "node-sass": "^4.11.0",
-    "popper": "^1.0.1",
+    "popper.js": "^1.15.0",
     "showdown": "^1.9.0",
     "ts-node": "^7.0.1",
     "tslint": "~5.11.0"
