@@ -7,8 +7,7 @@ import { FeedbackSentComponent } from './sent/sent.component';
 import { UtilsModule } from '../utils/utils.module';
 import { TranslateModule } from '@ngx-translate/core';
 import { FormsModule } from '@angular/forms';
-import { RecaptchaModule } from 'ng-recaptcha';
-import { RecaptchaFormsModule } from 'ng-recaptcha/forms';
+import {RecaptchaFormsModule, RecaptchaModule} from 'ng-recaptcha';
 
 @NgModule({
   declarations: [FeedbackComponent, FeedbackSentComponent],
