@@ -6,8 +6,7 @@ import { RestorePasswordComponent } from './restore-password.component';
 import { RestorePasswordSentComponent } from './sent/sent.component';
 import { RestorePasswordNewComponent } from './new/new.component';
 import { RestorePasswordNewOkComponent } from './new/ok/ok.component';
-import { RecaptchaModule } from 'ng-recaptcha';
-import { RecaptchaFormsModule } from 'ng-recaptcha/forms';
+import {RecaptchaFormsModule, RecaptchaModule} from 'ng-recaptcha';
 import { UtilsModule } from '../utils/utils.module';
 import { FormsModule } from '@angular/forms';
 import { TranslateModule } from '@ngx-translate/core';
