@@ -4,8 +4,7 @@ import { CommonModule } from '@angular/common';
 import { SignupRoutingModule } from './signup-routing.module';
 import { SignupOkComponent } from './ok/ok.component';
 import { SignupComponent } from './signup.component';
-import { RecaptchaModule } from 'ng-recaptcha';
-import { RecaptchaFormsModule } from 'ng-recaptcha/forms';
+import {RecaptchaFormsModule, RecaptchaModule} from 'ng-recaptcha';
 import { UtilsModule } from '../utils/utils.module';
 import { FormsModule } from '@angular/forms';
 import { TranslateModule } from '@ngx-translate/core';
